requirements.txt
flask
requests
pymongo
pyyaml